fix: change records_tasks fields from Integer to BigDecimal to match db-schema.md
- Liquibase v1.0.4-1 alters 5 count columns to DECIMAL(10,2)
- RecordsTask entity, RecordsDto, RecordsUpdateRequest updated to BigDecimal
- RecordsService defaults updated to BigDecimal.ZERO
- JacksonConfig strips trailing zeros so DECIMAL(10,2) serializes as plain numbers
- Fix TravelControllerIT matchers from 60.0f/40.0f to 60/40 (now plain ints in JSON)
- Fix RepositoryConstraintTest and RecordsServiceTest to use BigDecimal values

+++ backend/src/main/java/com/workload/dto/RecordsDto.java
@@ -1,12 +1,13 @@
 package com.workload.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record RecordsDto(
     UUID id,
     UUID objectId,
-    Integer accessRequests,
-    Integer monitoringRequests,
-    Integer footageRequests,
-    Integer backupControl,
-    Integer securityAdmin) {}
+    BigDecimal accessRequests,
+    BigDecimal monitoringRequests,
+    BigDecimal footageRequests,
+    BigDecimal backupControl,
+    BigDecimal securityAdmin) {}
